test: Clean up Fan.IntegrationTests before Widget development #277
